requirements.txt
pyRFXtrx==0.25.0
paho-mqtt==1.5.0
PyYAML==5.3.1
